rewrote modules to be split into separate code
