+++ src/Symengine/VecBasic.hs
@@ -1,0 +1,100 @@
+{-# LANGUAGE RecordWildCards #-}
+{-# LANGUAGE InstanceSigs #-}
+{-# LANGUAGE FunctionalDependencies #-}
+module Symengine.VecBasic
+  (
+    VecBasic,
+    vecbasic_new,
+    vecbasic_push_back,
+    vecbasic_get,
+    vecbasic_size,
+    list_to_vecbasic,
+   )
+where
+
+
+import Prelude
+import Foreign.C.Types
+import Foreign.Ptr
+import Foreign.C.String
+import Foreign.Storable
+import Foreign.Marshal.Array
+import Foreign.Marshal.Alloc
+import Foreign.ForeignPtr
+import Control.Applicative
+import Control.Monad -- for foldM
+import System.IO.Unsafe
+import Control.Monad
+import GHC.Real
+import Symengine
+
+import Symengine.Internal
+import Symengine.BasicSym
+
+-- |represents a symbol exported by SymEngine. create this using the functions
+-- 'zero', 'one', 'minus_one', 'e', 'im', 'rational', 'complex', and also by
+-- constructing a number and converting it to a Symbol
+-- 
+-- >>> 3.5 :: BasicSym
+-- 7/2
+--
+-- >>> rational 2 10
+-- 1 /5
+--
+-- >>> complex 1 2
+-- 1 + 2*I
+-- vectors binding
+
+-- | Represents a Vector of BasicSym
+-- | usually, end-users are not expected to interact directly with VecBasic
+-- | this should at some point be moved to Symengine.Internal
+newtype VecBasic = VecBasic (ForeignPtr CVecBasic)
+
+instance Wrapped VecBasic CVecBasic where
+    with (VecBasic p) f = withForeignPtr p f
+
+
+-- | push back an element into a vector
+vecbasic_push_back :: VecBasic -> BasicSym -> IO ()
+vecbasic_push_back vec sym =  with2 vec sym (\v p ->vecbasic_push_back_ffi v p)
+
+
+-- | get the i'th element out of a vecbasic
+vecbasic_get :: VecBasic -> Int -> Either SymengineException BasicSym
+vecbasic_get vec i =
+  if i >= 0 && i < vecbasic_size vec
+  then 
+    unsafePerformIO $ do
+    sym <- basicsym_new
+    exception <- cIntToEnum <$> with2 vec sym (\v s -> vecbasic_get_ffi v i s)
+    case exception of
+      NoException -> return (Right sym)
+      _ -> return (Left exception)
+  else
+    Left RuntimeError
+
+
+-- | Create a new VecBasic
+vecbasic_new :: IO VecBasic
+vecbasic_new = do
+    ptr <- vecbasic_new_ffi
+    finalized <- newForeignPtr vecbasic_free_ffi ptr
+    return $ VecBasic (finalized)
+
+
+list_to_vecbasic :: [BasicSym] -> IO VecBasic
+list_to_vecbasic syms = do
+  vec <- vecbasic_new
+  forM_ syms (\s -> vecbasic_push_back vec s)
+  return vec
+
+vecbasic_size :: VecBasic -> Int
+vecbasic_size vec = unsafePerformIO $
+  fromIntegral <$> with vec vecbasic_size_ffi
+
+foreign import ccall "symengine/cwrapper.h vecbasic_new" vecbasic_new_ffi :: IO (Ptr CVecBasic)
+foreign import ccall "symengine/cwrapper.h vecbasic_push_back" vecbasic_push_back_ffi :: Ptr CVecBasic -> Ptr CBasicSym -> IO ()
+foreign import ccall "symengine/cwrapper.h vecbasic_get" vecbasic_get_ffi :: Ptr CVecBasic -> Int -> Ptr CBasicSym -> IO CInt
+foreign import ccall "symengine/cwrapper.h vecbasic_size" vecbasic_size_ffi :: Ptr CVecBasic -> IO CSize
+foreign import ccall "symengine/cwrapper.h &vecbasic_free" vecbasic_free_ffi :: FunPtr (Ptr CVecBasic -> IO ())
+
