+++ symengine.cabal
@@ -15,7 +15,12 @@ cabal-version:       >=1.10
 
 library
   hs-source-dirs:      src
-  exposed-modules:     Symengine
+  exposed-modules:     Symengine,
+                       Symengine.DenseMatrix,
+                       Symengine.BasicSym
+
+  other-modules:       Symengine.Internal,
+                       Symengine.VecBasic
   build-depends:       base >= 4.5.0 && <= 5
   default-language:    Haskell2010
